split github_services into seperate poros classes and services class
