+++ app/poros/pull_requests.rb
@@ -1,0 +1,6 @@
+class PullRequests
+  attr_reader :count
+  def initialize(data)
+    @count = data.count
+  end
+end
