+++ app/poros/repo.rb
@@ -1,0 +1,6 @@
+class Repo
+  attr_reader :name
+  def initialize(data)
+    @name = data[:name]
+  end
+end
